Split SvgTextPath
