+++ Source/Text/SvgTextPath.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-#if !NO_SDC
-using System.Drawing.Drawing2D;
-#endif
 
 namespace Svg
 {
@@ -51,25 +48,6 @@ namespace Svg
             get { return GetAttribute<Uri>("href", false); }
             set { Attributes["href"] = value; }
         }
-#if !NO_SDC
-
-        protected override GraphicsPath GetBaselinePath(ISvgRenderer renderer)
-        {
-            var path = this.OwnerDocument.IdManager.GetElementById(this.ReferencedPath) as SvgVisualElement;
-            if (path == null) return null;
-            var pathData = (GraphicsPath)path.Path(renderer).Clone();
-            if (path.Transforms != null && path.Transforms.Count > 0)
-                using (var matrix = path.Transforms.GetMatrix())
-                    pathData.Transform(matrix);
-            return pathData;
-        }
-        protected override float GetAuthorPathLength()
-        {
-            var path = this.OwnerDocument.IdManager.GetElementById(this.ReferencedPath) as SvgPath;
-            if (path == null) return 0;
-            return path.PathLength;
-        }
-#endif
 
         public override SvgElement DeepCopy()
         {
